CSN: CSN-leap-1
Initial cut at a Leap package, which extracts the non-modeling
components from the HAsim package and copies them here. This 
includes the build tools, non-modeling specific modules and 
submodels and the non-model demo designs. Changes include renaming
the build tools to leap-* from hasim-* and renaming the perl 
library to Leap:: from HAsim::. 

To use the Leap tools/modules one should check out the Leap package
and have a sufficiently new verion of the awb package installed
so that the leap-* tools are defined (in asimstarter). Finally,
the model (.apm file) must use the type 'Leap'.

To ease the transition from the modules in the HAsim package to
those in the Leap package, the modules and submodels continue to 
live in the the same relative (uniondir) directory (this includes 
the use of some hasim/ subdirectories in the Leap pacakge). That 
means that the uniondir mechanism will pick out the files from the
first package in the SEARCHPATH - preferably Leap. Checking out 
Leap after HAsim should result in the correct behavior.

Note further that all the modules and submodels in the Leap package
currently continue to exist in the HAsim package. This means that 
users who checkout HAsim and not Leap should be unaffected by 
changes in the Leap package. Soon, however, use of the HAsim package
without the Leap package will be deprecated as new functionality 
will only be added to the Leap tools and eventually redundant files
in the HAsim package will be deleted.