CSN: CSN-leap-26
In the SCons scripts, coerce commands in the hardware build
to be spawned under bash insteaad of sh. This allows the bsc 
log builder to work, even with the reference to the bash-only
$PIPESTATUS variable.

This eliminates what I believe to be the last reason for
requiring /bin/sh to be a reference to bash instead of dash.
This means that on Ubuntu systems, one doesn't need to
change the /bin/sh symbolic link to point at bash any more.