CSN: CSN-leap-34
Patch for the models broken by the previous check-in. This patch is
shaky in the sense the it special cases ROOT_DIR_MODEL in the
recursive SW configuration step in leap-configure.  The real solution
to this problem is probably moving the SW recursion into the build
pipeline, but that will take time.