CSN: CSN-leap-41
Reworked memory fifo to use the FIFOCount interface.  We were already
tracking count as part of implementation, but now it can be optionally
exposed to the instantiator.  This was needed for implementing
rudimentary flow control to work around the rrr sw-handle thread's
deadlock issues.  I really wish it had non-blcoking writes.