CSN: CSN-leap-85
Added a new function to retrieve build tree paths for files in
ModulList.  This could probably be a little more parametric.

Also, uppgraded the multi-fpga parser to the point that it can be used
in an apm as part of the multifpga build pipeline.  This pipeline
currently resides in a development repository so as not to pollute the
main line codes.