CSN: CSN-leap-120
Some major changes to the build pipeline.  Most of these changes are created by a desire to simplify dependency calculation for synthesis boundaries in SCons.  MultiFPGA has been using this style of build for about a month, and this could be considered a back port.

1) wrapper generation moved to build pipeline.  two wrapers are built,
one with static and one with dynamic sizing.

2) stub file spamming.  Synthesis boundaries used to only work if the
boundary was instantiated by its parent in the module hierarchy.  Now
they can be instantiated anywhere. We also differentiate this stub and
the wrapper file by adding a _compile string to the name of the
wrapper.  This helps avoid confusion in SCons.