CSN: CSN-leap-133
General work on vico compatibility.  Unfortunately, although vico is
quite close, some of the software (a static function relating to
determining whether a thread is the vico main thread seems to be
undefined) is still not compiling.  However, things do seem close.
Generally, the code was using non-awb style commits, and this caused
some obvious issues, particularly with regard to the awb/asim switch
in the include path.

I'm hoping for someone to examine the software issue, but for now, I
will try to hack around it and see how far I get.