CSN: CSN-leap-167
mkSizedBRAMFIFO() now takes the size of the FIFO as an integer parameter instead
of a NumTypeParam#() hack.  The code internally is ugly (taken from the new
Bluespec-supplied version) since there is no good way to convert an integer
value to a type.  However, the interface now matches mkSizedFIFO() and is easier
to use.