CSN: CSN-leap-195
Elliott realized that most of the traffic in a HAsim multi-FPGA configuration
is from passing an empty token back and forth in the global scratchpad ring.
This is not a good use of the interconnect (duh), especially since traffic is
infrequent enough on that ring that no arbitration is needed.

 - Generalize the token ring and interface into an addressable ring.
 - Add a mkConnectionAddrRingNode() with the same addressable properties
   as the token ring but no token.  When a node has data it sends a message.