CSN: CSN-leap-220
Memory services in the platform may now be compiled as a separate synthesis
boundary.  All that is needed is a "%param SYNTH_BOUNDARY mkMemServices" in
the AWB file.

It can't be turned on yet because there is no syntax in the multi-FPGA
code to describe synthesis regions that belong on all FPGAs.