CSN: CSN-leap-223
Add a global string table, managed by module context, as part of soft services.
The string table generates integer UIDs (currently 32 bits) for a given string.
Note: while UIDs are unique, there may be more than one UID mapped to a given
string.

To allocate a string, a module must be a member of the COLLECTED_MODULE context.
Allocate an entry with:
    GLOBAL_STRING_UID str <- getGlobalStringUID("foo");
getGlobalStringUID() looks for the string in the table.  If present, it returns
the existing UID.  If not present, a new UID is allocated and returned.

The global string table is built as a single database during compilation.
The database is loaded at run time by passing the path in the "-global-strings"
argument.  The standard workload run scripts do this automatically.

Both the synthesis boundary ID and the platform ID are components of a
string's UID, allowing for parallel allocation of global string UIDs during
compilation.