CSN: CSN-leap-227
Enhanced leap-configure to pick up its options using the
standard awb workspace configuration files (awb.config,
~/.asim/asimrc and <prefix>/etc/asim/asimrc. Invoke the
following to see the options:

   % leap-configure options

Also converted leap-configure to take action commands, such
as 'configure' and 'run' rather than actions switches, such
as -configure and -run, to make it consistent with the other
configuration programs, such as amc(1). The old form still
works and is used by the Perl library, which needs to get
upgraded. But the prefered use of leap-configure is now:

   % leap-configure configure --model=config/pm/.../foo.apm

Also made leap-configure exit more gracefully in the face of 
bad model parameters and added documentation accessable with:

   % leap-configure help

Finally, made bluesim builds work across differnt versions
of gcc due to difference is the availability of the 
no-var-tracking-assignments option. This is accomplished by a
new utility function getGccVersion(). This closes bug #100.