CSN: CSN-leap-232
Debug scanning now uses global strings instead of dictionary entries.  Clients
may define fields within a scan record that are parsed and printed in
human-readable form in the software.  There is no hardware overhead for the
parsing, since the formatting descriptor is stored in the global string table.

Functions are provided in the debug scan service for constructing descriptors.