CSN: CSN-leap-260
Restructure the debug scan pipeline slightly so that the master FPGA is on a
private ring.  The hardware side scanner first scans the master ring and then
scans a global ring of all slave FPGAs.  In theory this will give us dumps on
the master even when some inter-FPGA router is deadlocked.