CSN: CSN-leap-263
Usually in the multiFPGA router synthesis, we try to be as efficient
as possible in terms of the number of bits transported.  However, in
the case that we generate stats instrumentation, the way that we
calculate these bitwidth can be wrong, and lead to wrong programs.  We
have a plan to fix the caclulation issue, but until we do, I will for
chains to have the maximum bitwidth.