CSN: CSN-leap-273
Added a new physical connection type.  Internally, some modules of the
multifpga router need to know if it is safe to schedule an access to
the shared router buffer.  To do this, they need to track buffer
availability.  As a result, they need to see dequeue operations from
the soft connection components.  This check in adds the necessary
visibility.  Existing SC implementations will see the original SC
interface.