CSN: CSN-leap-297
Replace our topological build with a build of all synthesis boundaries in parallel
followed by a final "link" step in which all boundaries are included by the
top level model build.

In the old scheme each synthesis boundary built a wrapper module that was
the true boundary compilation and a stub module that included the exposed
wires from the boundary.  The parent was required to instantiate the stub.

The new scheme is similar.  There is still a wrapper with the true compilation
and a stub.  However, the stub is instantiated by the top level module
automatically by the build manager.  A dummy stub module that does nothing
is created by the build manager for two purposes:  compatibility with the
old scheme and so that parent modules may continue to act as though they
instantiate subordinate modules.  This also allows for writing code that will
work unchanged for declaring a module a synthesis boundary or not a boundary.

- BSV.py now creates the synthesis boundary stub in as <module_name>_synth.bsv
  instead of <module_name.bsv.  Also fixed a bug related to multiple targets
  in the top-level model build.

- Move management of multi-FPGA FPGA assignment from the wrapper gen code
  in the multi-FPGA library to standard Module.py.  Modules are now decorated
  with FPGA assignment in the Python SCons code.

- wrapper_gen.py instantiates all synthesis boundaries in a special module
  (instantiateAllSynthBoundaries).  The top level Bluespec model must instantiate
  this module because it must be done within the soft connections context.

- leap-connect now understands the new synthesis boundary compilation names and
  how to create dummy stubs.