CSN: CSN-leap-324
The expected compressed output type of a Compress typeclass is now a tagged
union, making it easier for the multi-FPGA compiler to manage channel
mapping.  For channel mapping within Bluespec, a new CompressionChunks
typeclass is now available for converting a tagged union to an HList of
chunks.  The existing CompressionMapping typeclass can then map the chunks
to I/O channels.