CSN: CSN-leap-328
Prepare leap package for using perlqt4 with conditional build
of leap-front-panel-qt3 or leap-front-panel (currently empty).
Control of build if provided with a --enable-perlqt4 switch
to ./configure. Default for that switch is 'no', but is 
automatically turned to yes for Ubuntu Oneiric and later.

Also removed build dependence on libperl-qt for debian package
builds, because libperl-qt is for qt3 and the package name
is different on qt4, and there is no way to choose a build
dependence conditionally.