CSN: CSN-leap-343
Add new memory interface mappers that present a multi-read port interface
to the client but maintain storage within a memory having a single read port.
There are two versions.  One maintains strict read/write ordering relative to
the times at which requests arrive.  The other triggers writes immediately.
The latter is more efficient as long as the client can deal with the
unpredictability.