CSN: CSN-leap-384
Add a new class of metadata passed through the LEAP memory system for read
requests:  RL_CACHE_GLOBAL_READ_META.  For now, the struct has only one
field that indicates whether the request is a prefetch.  Unlike the other
metadata field, which is opaque and has meaning only to the module that
generates it, the global metdata has the same definition everywhere.

The first use of this (not yet implemented) will be to permit lower levels
of the hierarchy to treat prefetch requests differently than normal reads.

The global read metadata travels with both read requests and responses.  One
could argue it has no use with responses, but the structure is small and it
is now easy to add other fields.