CSN: CSN-leap-387
New code implements dual write ports for 1-bit LUTRAMs using a pair of LUTRAMs.
Works with any type that can pack into Bit#(1).  Zero is encoded as both
identical values and one as different values for a given index.

The code works only if the same address is never passed to both write ports
in a given cycle.  This is not checked as a check would add scheduling
complexity and, in some cases, introduces a dependence between attempts
to use both write ports in parallel.