CSN: CSN-leap-422
1) Refactored run.bmark to support multiFPGA builds.  run.bmark now
interoggates FPGA sub-builds to find out which FPGAs are needed.  It
then programs all FPGAs in the system.  When the program step is done,
all FPGAs are activated, ensuring distributed reset. Finally, the
software side program is executed.

The new functionality supports only FPGAs attached to the same system.
Remote FPGA support is coming soon, as is support for multiple
software instances.