CSN: CSN-leap-444
Resurrected support for the sub-build version of synplify.  This
version uses synplify for the user portion of the program and xst for
the physical platform interface. 

I also factored out some common functions used in both the synplify
and xst builds into a common library.

A warning - although the synplify build produces bitfiles, these
bitfiles appear to have some unresolved timing issues. As such, the
synplify build is somewhat unreliable at this time.