CSN: CSN-leap-468
Add a method to counters that returns the value of the counter at the end of
the cycle, after updates have been applied.  This is in addition to the
current value() method that returns the counter value at the beginning of
the FPGA cycle.

Merge COUNTER and COUNTER_Z into the same interface and use a single
implementation.