CSN: CSN-leap-471
First step of making it possible to pass more configuration to a scratchpad
constructor.  A new SCRATCHPAD_CONFIG type is passed in to the constructor.
The type has a default value, so new parameters can be added.

In the next step I'll add the ability to initialize the scratchpad at run
time from a file.