CSN: CSN-leap-489
Another large update to leap in support of migration to the LIM compiler. 

1) Refactored RRR code generation to avoid referencing magic global
objects.  These references, if necessary, are hidden in the
implementation of the RRR client.  In the new code, this allows us to
avoid a call-through.

2) Added a signal handler for segfaults to generated top-level
header. This helps debug those code races which are perturbed by
valgrind.

3) Large update to li-channels library.  The code now automatically
matches sends and receives.  Sends are assumed to have context and
invoke methods on the received side.  This enables physical receives
to push data through I/O without relying on a seperate
thread. Dangling sends and receives can be resolved at link time for
those users of the LEAP shared library. 

4) Added new methods to physical channels to support LIM compiler.