CSN: CSN-leap-537
1) Fixed bug in route-through in which packet header was recoded
prematurely.

2) Added support for a route-through thread.  This thread is
instantiated per route-through (a terrible architecture, but fixable)

A better architecture involves a single thread and an out-of-order
send queue. Each channel is inorder, but if a send would block, it is
dequeued re-injected at the tail of the work queue. This enables a
single thread to handle all route-throughs (RRR can proably also be
handled this way also), reducing the thread overhead issues that we
currently face.