Added support for synplify premier to enable the use of systemverilog in user designs.
The chief code reorganization required by this was moving the XST LI
module synthesis routine to the shared synthesis library. Synplify
tends to change netlist names in ways that break ucf files.  Since the
platform code is now handled as an LI module, it needs the XST module
synthesis routine.