Added rudimentary area-group inference tool.
This tool attempts to place modules in an intelligent manner within a
single FPGA.

Also included is a parser allowing users to express white spaces and
platform-specific area groups within their designs.

This tool is integrated in to the LIM compiler as an experimental
feature.  It can be enabled with:

ENABLE_SMART_AREA_GROUPS=1

It is disabled by default, since it is currently unable to operate on
large designs.