Added rudimentary mechanism for importing pre-placed-and-routed design components across designs.  We can now specify these components as part of a .awb, and the backend flow will integrate them automatically.
Some improvements are needed.  Ideally, we would have a database of
these components, and they would be visible to the area group
placement algorithm. Since device codes are manually placed, we don't
have that problem yet.  Additionally, we need some check (like a hash
of the original verilog) to ensure that the result we are getting is
actually useable in the current design.