Switch to a new strategy for annotating sync FIFOs based on advice from http://forums.xilinx.com/t5/Timing-Analysis/Timing-Failure-MCMM-with-multiple-outputs/td-p/563409
We now discover the two clocks in a sync FIFO automatically and then tag
the two clock domains using clock groups.  Instead of attempting to decorate
all cross-clock paths and suffering the consequences as optimization changes
names, merges cells, etc. we now simply assert that the two clocks are
asynchronous and that all paths appearing to cross the clock domain are managed
correctly by the code.