Added an access tracker in the set associative BRAM cache.
The tracker tracks the number of accesses to each line during its
lifetime in the cache, up to some saturating value.  Upon eviction,
the number of accesses is written to a histogram.  The variable
RL_CACHE_LINE_ACCESS_TRACKER_WIDTH sets the width of the histogram,
which is 0 and disabled by default.