Added some support for non-power-of-two sized DM caches.
We achieve non-power of two by load-balacing the high-order address
bits across the cache.  For example, we might map four bits to two
bits to achieve a 3/4 power of two cache, such that each 1/4 of the
cache gets ~16/3 of the accesses.

Both the direct mapped cache and scratchpad can select this cache.